Determine inclusion pointwise and plot true contour dotted
Inclusion of CoPE sets is now determined pointwise and no longer with the function SubsetContour as this was giving wrong results for large sample sizes.
The function SubsetContour is no longer necessary and removed.

If the true function is available the true contour is now plotted as a dotted purple line and the estimated contour is always plotted as a solid line.

The inverse square root of the correlation matrix that we need in ComputeCope can be extracted directly from the nlme package and we do that now.